Create my own implementation of jlox
It follows the Java implementation of Lox given in the book
"Crafting Interpreters" (https://www.craftinginterpreters.com/).
But one major difference is that, rather than generating the code
for the `Expr` and `Stmt` classes using string manipulation as the
book did, I used a Java library called JavaPoet to generate them
instead.